Changed api to number regex. Configured $http in run()
The API should have primitive versioning and regex handling for number
parameters.

Angular should be configured in the run() function so that the $http
service uses the usual urlencoded data instead of json encoded ones.
Simplifies the $http methods so that $http.<method> can be called
instead of configuring everything each time.